requirements.txt
memory-profiler==0.61.0
orjson==3.10.6
google-cloud==0.34.0
google-auth==2.32.0
emoji==2.12.1
snakeviz==2.2.0